+++ astarte-utils/AstarteGenericProducer.cpp
@@ -127,6 +127,18 @@ bool AstarteGenericProducer::sendData(const QVariant &value, const QByteArray &t
     return false;
 }
 
+bool AstarteGenericProducer::sendData(const QList<QVariant> &value, const QByteArray &target, const QDateTime &timestamp, const QVariantHash &metadata)
+{
+    // TODO: checks
+    QHash<QByteArray, QByteArray> attributes;
+    attributes.insert("interfaceType", QByteArray::number(static_cast<int>(m_interfaceType)));
+    // TODO: handle reliability, retention and expiry
+
+    sendDataOnEndpoint(value, target, attributes, timestamp, metadata);
+
+    return true;
+}
+
 bool AstarteGenericProducer::sendData(const QVariantHash &value, const QByteArray &target, const QDateTime &timestamp, const QVariantHash &metadata)
 {
     // TODO: verify path match